refactor: split dashboard API module by responsibility
