+++ frontend/src/api/dashboard/index.ts
@@ -1,0 +1,33 @@
+export type {
+  BreakdownCategoryKind,
+  CategoryBreakdownEntry,
+  CreditWidgetResponse,
+  FxRateIssue,
+  FxState,
+  FxStatus,
+  MonthlyIncomeExpense,
+  NetWorthWidgetResponse,
+  RecentActivityWidgetResponse,
+  SavingsRateWidgetResponse,
+  SpendingBreakdownResponse,
+  SpendingComparisonResponse,
+  SpendingRange,
+} from '@/api/dashboard/types';
+
+export {
+  fetchDashboardCredit,
+  fetchDashboardNetWorth,
+  fetchDashboardRecentActivity,
+  fetchDashboardSavingsRate,
+  fetchSpendingBreakdown,
+  fetchSpendingComparison,
+} from '@/api/dashboard/requests';
+
+export {
+  useDashboardCredit,
+  useDashboardNetWorth,
+  useDashboardRecentActivity,
+  useDashboardSavingsRate,
+  useSpendingBreakdown,
+  useSpendingComparison,
+} from '@/api/dashboard/hooks';
